[open][neva_upstreamable][webrtc] Suspend video capture when app is not focused.
:Release Notes:
Suspend video capture when app is not focused.

:Detailed Notes:
If application is not focused, video capture should stop, to avoid
broadcast by mistake users intimacy.

This patch reuses Android implementation for same purpose. It does NOT
stop audio broadcasting, as it is considered interesting to be able to
keep a conversation while consuming other device contents. But it would
be desirable to integrate this feature in UI (I.e. with a media capture
notification).

:Testing Performed:
1. Open Skype
2. Do a video call: video is broadcasted to the other party.
3. Change app: video is paused in the other side.
4. Go back to video call: video recovers and continues broadcast.

:QA Notes:
It does not suspend audio broadcast.
Test-case: N/A

:Issues Addressed:
[PLAT-42746] WebRTC: suspend/resume support
[PLAT-65882] Chromium 2018Q4 upgrade: migrate "webrtc" patchset

Change-Id: I221a9400b32077f13ac518d25223003cb5fead2c
Originally-Reviewed-on: http://gpro.lgsvl.com/177294
Originally-Reviewed-on: http://gpro.lgsvl.com/229893
Reviewed-on: http://gpro.lgsvl.com/230152
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>