[BACKPORT][wayland gpu] Place drm format helper methods into shared place.
This CL is a part of Wayland gpu/ui split effort, which fixes the
Ozone/Wayland and makes it functional when the gpu service is running
off the browser process.

The patch does not bring any functionality changes, but rather
prepares Ozone/Wayland dmabuf based approach, which uses ozone/drm code,
to be landed in the upstream.

In particular, drm_util is partly moved to ui/ozone/common/linux and
implements drm format conversion methods, which are used by Ozone/drm
and are going to be used by Wayland with dmabuf approach.

Bug: 820047
Change-Id: I024fd1f6a57b9e9693864e62d5cb8a443afd9e2a
Reviewed-on: https://chromium-review.googlesource.com/1071792
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#561977}