fixup! [open][neva_upstreamable][webos_webview] Implement SetKeepAliveWebApp
:Release Notes:
Ensure renderer preferences keep alive flag is set immediately in
renderer process after setting from injection.

:Detailed Notes:
This patch is related to a crash problem(PLAT-60609) which has reproducibility about 50%~99%
This patch is a sort of code for safety eventhough it's not easy to be tested to prove this issue as of now.

:Testing Performed:
Youtube(KeepAlive()), play video, press back till the close app dialog is shown
and close. Do it 2 more times. Now it works. Without the patch the third
time it would stop rendering.

:QA Notes:
Test-cases: GT-25234

:Issues Addressed:
[PLAT-69241] Chromium 2018Q4 upgrade: migrate "wam" patchset - Patchset 188..260 - 2 188..260 - 2

Change-Id: Ic75f3d00b08695560a9bbaec2a4f4501c78ee4da
Originally-Reviewed-on: https://gpro.lgsvl.com/209240
Originally-Reviewed-on: https://gpro.lgsvl.com/221189
Originally-Reviewed-on: http://gpro.lgsvl.com/224965
Reviewed-on: http://gpro.lgsvl.com/225300
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>