[BACKPORT][wayland gpu] Add frame callback support.
Instead of just assuming buffer is committed, use wayland provided
frame callbacks.

Once the GbmSurfaclessWayland (resides on the GPU process side)
requests to swap buffers from Wayland (the connection is established
on the browser process side), it sends a ScheduleBufferSwap requests
and provides a completion callback, which is then triggered once
Wayland sends a frame callback (setup for each new buffer swap).

This callback is then used to identify if the swap was successful
or not.

Bug: 859012
Change-Id: I3264db6bb4456ef610d11945199ecddffbc23024
Reviewed-on: https://chromium-review.googlesource.com/1193882
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Robert Kroeger <rjkroege@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Cr-Commit-Position: refs/heads/master@{#589457}