[open][upstreamable][build] Allow to pass extra build flags to clang toolchains.
:Release Notes:
Allow to pass extra build flags to clang toolchains.

:Detailed Notes:
On using clang toolchain, the build will use the self-contained one
included in the build. But it does not accept extra build flags as
it was possible in GCC.

This patch just allows to pass extra flags, as the ones we could
set in yocto specified toolchains.

:Testing Performed:

:QA Notes:

:Issues Addressed:
[PLAT-66448] Chromium 2018Q4 upgrade: pkg-config-host hack fetches nss from desktop

Change-Id: I5103e8385ffc6edff8ad8f66443fc89ea3e4603a
Originally-Reviewed-on: http://gpro.lgsvl.com/217885
Reviewed-on: http://gpro.lgsvl.com/218537
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>