[open][neva_upstreamable][cbe] CBE build
:Release Notes:
Make CBE build for app_shell and webOS webview.

:Detailed Notes:
Build Chromium stack as a single library (libcbe.so) that includes
everything but the executable for app_shell. It also adds support
for including wam_demo.

It adds the structure needed for adding other executables later
(as vanilla Chromium, content_shell or chromedriver).

The most complex part has been working around the testonly checks.
In upstream chromium, app_shell has testonly = true. As we expect
app_shell to not be a test binary, we had to provide an alternate
for the content_shell_lib that is not testonly. It is a hack and
eventually we should propose upstream a way to build app_shell
as not testonly while still keeping content_shell as is (testonly,
and with layout test infrastructure available).

It replaces these original patches in chromium53:
https://gpro.lgsvl.com/#/c/130962/
    Author: Vladislav Mukulov <vladislav.mukulov@lge.com>
    Date:   Wed Dec 23 09:44:02 2015 +0300
    [cbe] Make CBE build for chromium
https://gpro.lgsvl.com/137428
    Author: Philipp Smunev <philipp.smunev@lge.com>
    Date:   Wed Mar 11 19:17:11 2015 +0300
    [cbe] Load chromium plugins from the cbe library folder
https://gpro.lgsvl.com/157260
    Author: Jose Dapena Paz <jose.dapena@lge.com>
    Date:   Thu Nov 3 11:31:13 2016 +0100
    [cbe] Install and use locales and resources under CBE path
https://gpro.lgsvl.com/145447
    Author: Jose Dapena Paz <jose.dapena@lge.com>
    Date:   Mon Jun 13 15:55:32 2016 +0200
    [webos_webview][chromium44][cbe] Add shared path for CBE shared data files

:Testing Performed:
Ran on TV with app_shell browser and WAM. The binary for app_shell
is below 10k, and depends on libcbe.so (below 100MB). If wam_demo is enabled
it is also around 10-15k.

:QA Notes:
Test cases: GT-25930
Depends on https://gpro.lgsvl.com/#/c/218872/

:Issues Addressed:
[PLAT-65789] Chromium 2018Q4 upgrade: migrate "initial" patchset

Change-Id: I4597df58591bc1bb544d522078ecf12680434795
Originally-Reviewed-on: http://gpro.lgsvl.com/218871
Reviewed-on: http://gpro.lgsvl.com/219346
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>