[open][upstreamable][input][ime] Support 'sensitive' attribute
:Release Notes:
Support 'sensitive' attribute on html input elements for protecting
the privacy information such as id, password or email from Voice
Recognition(STT).

:Detailed Notes:
The value of sensitive attribute is updated whenever input text
is changed. If the value is true, it means STT is enabled and
the information can't be written with Voice Recognition.

This patch is migrated from 'chromium53' project and originally
committed by Jeongran Jang on https://gpro.lgsvl.com/155440

:Testing Performed:
1. Host test_for_sensitiveAttr.html webpage listed in DRD-12220.
2. Three text input fields will be visible with different sensitive
   attributes.
3. Use mic to enter text after selecting STT option in VKB
4. Observe STT option will not be available for input text fields with
   sensitive attributes, first and second text box for the test page.

:QA Notes:
Test-cases: GT-20003

:Issues Addressed:
[PLAT-67282] Chromium 2018Q4 upgrade: migrate new attribute 'sensitive'
             on input element

Change-Id: I00f789642b190cf0fa33e3796e545022eb774019
Originally-Reviewed-on: http://gpro.lgsvl.com/219019
Reviewed-on: http://gpro.lgsvl.com/219349
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>