[open][upstreamable][fonts] Enforce fallback fonts to unicode script languages
:Release Notes:
Fix the problem that displays wrong fonts
On resolving fonts, guess language of the unicode script and add
it to the fontconfig query
Apply unicode range language guess always and ignore web page
language on calculating font fallback

:Detailed Notes:
1. There exist a missing logic to deliver the locale information
to the skia port FontCache. Than the blink requests to the
fontconfig library with 'empty' locale and it generates unexpected
ordered.(It means the language priority is ignored.)
This patch is inspired by the android port FontCache
that uses the 'defaultLanguage' when the documentStyle doesn't have
specific locale information
2. Our fontconfig settings contain rules for mapping languages to
certain fonts as fallback. This is used for indian fonts.
On WebKit GTK+ (with pango) and on QtWebKit (with Qt) the fontconfig
queries were hinted converting the unicode character range to a
language. But on Chrome this hinting did not exist, so for texts
without explicit language hint, or using fonts different from
current system font, the resolution could be wrong.
This patch implements the mapping from unicode ranges to languages,
and adds this hinting on resolving the font fallbacks.
3. On TV we have many apps that are providing a language through CSS.
But this was still not enough, and many times the system fallback
was used. So, on the fallback font, we just always override the
language with the unicode range language guess. This is because we
already failed looking for the font and we are already looking for
a fallback.
This makes the fallback more effective, specially on LG store app, that
was still showing Lohit fonts sometimes.

:Testing Performed:
1. Set locale: ml-IN
2. Go to Setting->General->Menu Langauge
3. Check dropdown listbox
4. LG Display for Indian fonts is OK (i.e. as-IN, bn-IN, gu-IN,
ml-IN, pa-IN, ta-IN)

:QA Notes:
Test-cases: GT-22206

:Issues Addressed:
[PLAT-66198] Chromium 2018Q4 upgrade: migrate "wam" patchset - fonts

Change-Id: I52dc5174c630be2a5fd7f9b3a47dd0ea5c5c30b7
Originally-Reviewed-on: https://gpro.lgsvl.com/158390
Originally-Reviewed-on: http://gpro.lgsvl.com/220389
Reviewed-on: http://gpro.lgsvl.com/221453
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>