[open][neva_upstreamable][webos] Introduce RuntimeDelegateWebOS in the content layer
:Release Notes:
To make content layer to use WebOS's runtime, we need to support
an interface so that the content layer can use WebOS's runtime resource.

:Detailed Notes:
This patch introduces |RuntimeDelegateWebOS| delegate interface, and
|WebOSRuntimeDelegate| inherits and implement it. And then, we create
the |WebOSRuntimeDelegate| and register it to the content layer at the
initialization stage.

:Testing Performed:
Chromium68 build was fine.

:QA Notes:
Test-cases: N/A
This patch is not to add any functionality

:Issues Addressed:
[PLAT-65860] Chromium 2018Q4 upgrade: migrate "a11y" patchset

Change-Id: Ia9fba5a456bf3c0b2f08b19a01b3904aa0b965b7
Originally-Reviewed-on: http://gpro.lgsvl.com/220631
Reviewed-on: http://gpro.lgsvl.com/220670
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>