[open][neva_upstreamable][network][locale] Reload resources on render process
:Release Notes:
Add support to reload locale on render process when running with
zygote mode on.

:Detailed Notes:
Add support to reload locale on render process when running with zygote
mode on. When running on zygote mode the ResourceBundle is intitialized
for zygote process and is shared across all renderer process without
creating new resource bundle.
Zygote process is initialized and launched before we get any locale
information from platform. Render process is notified about the language
change and reload the locale resources for the render process.
Also added check in ResourceBundle to avoid multiple reloads when
the notification is sent for each of the applications.

:Testing Performed:
Checked that locale changes are handled in the renderer process.
This is a part of needed changes to support reload locale.

:QA Notes:
Test-cases: GT-23562

:Issues Addressed:
[PLAT-66523] Chromium 2018Q4 upgrade: migrate "network" patchset - locale

Change-Id: Id3eba698edc705192006af2c48ec91088f573e12
Originally-Reviewed-on: https://gpro.lgsvl.com/177851
Originally-Reviewed-on: http://gpro.lgsvl.com/220854
Reviewed-on: http://gpro.lgsvl.com/221455
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Jani Hautakangas <jani.hautakangas@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>