[open][neva_upstreamable][ozone] Keep size of size-overriding apps
:Release Notes:
Keep size of size-overriding apps.

:Detailed Notes:
Some apps have a hardcoded size, as establiedh in appinfo.json. This is
passed to app runtime API. But on relaunching, the wayland window (and
app size) was resized to display screen size, after fullscreen.

There were two problems. One was that on setting the window fullscreen,
the bounds were set to the size of the display. This has been fixed
implementing the ToggleFullscreenWithSize method that was already
present in chromium53.

The other problem was that, once application got fullscreen status,
window manager was sending a configure event with the display size.
Solution is, as in chromium53, ignoring this. For that this time
we use the ui::WidgetDelegate::CanResize property, that is set
to false in app runtime already. Additionally this patch adds the
handling of the resize behavior that is populated using CanResize
value.

:Testing Performed:
Launch viewster (view is shown properly, taking all the screen). Then
switch to another application, and back to viewster. With this patch
the application keeps showing fullscreen.

:QA Notes:
Test-cases: GT-15269

:Issues Addressed:
[PLAT-70480] [chromium 68] Wrong size of apps after switching via icon

Change-Id: I93189de67e4aff2a3c37da07d29851cea4ebf79e
Originally-Reviewed-on: http://gpro.lgsvl.com/224298
Reviewed-on: http://gpro.lgsvl.com/224492
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>