Revert "[open][neva_upstreamable][splash] Sync FMP and swap for FMP to dismiss splash"
:Release Notes:
This reverts commit 46030d87a494123c0f60f2639a5fc0b8d54a529a.

:Detailed Notes:
The patch is not needed anymore because chromium68
FirstMeaningfulPaintDetector has been improved to get the time of
next swap for FMP.
So it is enough to rely on FMP itself

:Testing Performed:
Compare app launch performance for CP and system apps side by side with
master build(v53) - should be similar
Result: Performance is very close to v53. There is still some slowness on system
apps like photo-video, music, which can be tuned further in next set of patches.

:QA Notes:
Test-cases: GT-18470
Depends on: http://gpro.lgsvl.com/224437

:Issues Addressed:
[PLAT-69598] [chromium68] Poor app launch performance

Change-Id: Ia94eeaf1ff280e58c4abcf7258af85d0e6615d55
Originally-Reviewed-on: http://gpro.lgsvl.com/224436
Reviewed-on: http://gpro.lgsvl.com/224834
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>