[open][downstream][media] Provide additional contents scale to Media stack
:Release Notes:
Upscale video to fullscreen providin additional contents scale to media
stack.

:Detailed Notes:
Playing video, it is needed to provide the target position for
the video in UI resolution (and indicate if it is fullscreen or not). For
that, the algorithm uses additional contents scale, to adapt the media
resolution to the UI resolution.

This is needed for properly showing fullscreen videos on videos that
override the application resolution at fullscreen (regular way to see apps
in webOS). The window will get the overriden resolution, but it is
upscaled to show fullscreen. Videos need to adapt to the expected
resolution.

:Testing Performed:
Launch Viewster or any app which is not screen size (app with resolution 1280x720)
Play embedded or fullscreen video
Confirm video scaling & position is correct

:QA Notes:
Test-cases: GT-15269, GT-21876, GT-19638

:Issues Addressed:
[PLAT-69164] [chromium 68] Viewster CP app (incorrect video size)

Change-Id: I6d35305f228c614b50aa39c434a071df3a4734ba
Originally-Reviewed-on: http://gpro.lgsvl.com/224426
Reviewed-on: http://gpro.lgsvl.com/224493
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>