[agl] make SystemHostResolverProc to honor all the errors.
For some unknown reason SystemHostResolverProc honors only
EAI_SYSTEM errors and allows to other errors to pass by, which
results in invalid addrinfo and seg 11.

In upstream, it has always been like that and the last patch, which
touched that code dates back to 2010.
https://codereview.chromium.org/2134004

Thus, to make it work, honor all the errors and return
ERR_NAME_NOT_RESOLVED.