[agl] Use wl_surface_frame to send new frames to Compositor.
Previously, the DisplayCompositor has been sending new frames
regardless of the fact that a Wayland compositor might not
be ready to take more frames, or it has been about to go
background and would not be able to process frames until it was
in foreground.

It was ok when running only one application, but as soon as
a second app was opened, it couldn't connect to the GPU, because
it was blocked in eglSwapBuffers, and everything just hang.

Thus, use frame callbacks to be sure the frames are sent on time.

The wl_surface_frame has already been used with WaylandBufferManager,
but that class is only used with a separate GPU process.

This CL fixes --in-process-gpu mode.

SPEC-2080

* This change may be upstreamable.