[open][neva_upstreamable][app_shell] Let app shell ack webOS shell window close event
:Release Notes:
Acknowledge in app shell the webOS shell window close event to close windows.

:Detailed Notes:
webOS shell notifies when a window should be closed, and that is properly
handled already through the wayland window tree host for neva app runtime.

But we need to also acknoledge the event from app shell, to properly close
the window when the window manager asks.

:Testing Performed:
On Rpi build:
1. Launch enact browser app and open www.w3.org webpage
2. Launch youtube app and Play a media
As a result, the selected Video Content is not displayed,
instead previously loaded browser page (www.w3.org) is displayed.

:QA Notes:
Test-cases: QR-1364

:Issues Addressed:
[PLAT-72620] Selected Video Content is not displayed,
   instead Previously loaded browser page is displayed.
   But, Audio is heard.

Change-Id: I491179cd0b16b45810d5883943b35226b8906bbc
Originally-Reviewed-on: http://gpro.lgsvl.com/228126
Reviewed-on: http://gpro.lgsvl.com/228128
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>