[ose][downstream][media][ls2] Match URIs array with URITypes
:Release Notes:
Match URIs array with URITypes

:Detailed Notes:
luna_service_uris array is indexed based on URIType, but there was
a mismatch between the sizes of each.

As some services are not available in OSE, use empty strings so that
there's no buffer overrun when accessing it based on any URIType.

:Testing Performed:
Basic testing of OSE BAT.

:QA Notes:
N/A

:Issues Addressed:
[PLAT-72650] [OSE] [chromium68] Static analysis issues

Change-Id: Ife03dba34b574941a9851c7223d57b6b2f180ecd
Reviewed-on: http://gpro.lgsvl.com/228141
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>