fixup! [open][neva_upstreamable][media][ums] Introduce SystemMediaManager
:Release Notes:
Fix audio is not out sometimes

:Detailed Notes:
This patch has following changes to fix the issue
1. Notify modified PlayState when suspend and resume
By refactoring system media manager integration, the state of app and
player is sending as it is, but system media manager react
unexpectedly in such case of notifying
Background/Paused -> Foreground/Paused. By this state transition,
sometimes audio is not connected properly.
In Previous implementation, always Loaded state notified in
suspend/resume case instead of Paused if media-pipeline is loaded.

2. Also |SetPlaybackVolume| is added on |DispatchingPlaying| as before
to make sure that system media manager can connect the audio sink.

3. Added |SourceInfoUpdated| to SystemMediaManager interface.
It will help for SystemMediaManager to know audio/video availability
on time.

:Testing Performed:
On device QS-2506 is tested and audio was out well.

:QA Notes:
Test-cases: QS-2506

:Issues Addressed:
[PLAT-72634] Audio should be heard when video is played.
[PLAT-72828] Audio is not heard when USB video is played

Change-Id: I4212897968cfd700f0e88db8ae23796db55c0107
Originally-Reviewed-on: http://gpro.lgsvl.com/228513
Reviewed-on: http://gpro.lgsvl.com/228615
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>