[open][neva_upstream][media] Fixed sometimes youtube video shows black
:Release Notes:
Fixed sometimes youtube video shows black screen

:Detailed Notes:
|WebMediaPlayerMSE| doesn't initialize the |render_mode_| and it caused
undefined behavior such as skipping |SetRenderMode|.
If the uninitialized |render_medoe_| has 1 (RenderMode::RenderModeHole)
then |SetRenderMode| will skip calling |SetStorageType|.
The |storage_type_| of the |VideoFrameProviderImpl| remained as
media::VideoFrame::STORAGE_UNKNOWN and it failed to create the
VideoFrame. Finally |VideoFrameProviderImpl| has a nullptr for
|current_frame_|. Video was drawn as black color by this reason.
This patch adds the initialization of |render_mode_|

:Testing Performed:
 1. Launch the HTML5 Video from the launcher
 2. Run following commnad in web inspector console
   window.location = "http://www.youtube.com"
 3. Play any video
 4. If the video plays properly, play another video.

:QA Notes:
The probability of the issue is about under 5%. If the problem is
not seen then reload the site and try to play a video again. I was able
to see the problem within 10 minutes.
Test-cases: GT-26802

:Issues Addressed:
[NEVA-3113] Sometimes video does not play (using run_app_shell)

Change-Id: I4cff7c91c5b74d9d98fb2e72d6d24bf435562c6a
Originally-Reviewed-on: http://repo.lge.com:8080/1731
Originally-Reviewed-on: http://gpro.lgsvl.com/228925
Reviewed-on: http://gpro.lgsvl.com/229001
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>