[open][neva_upstreamable][cbe] Do not export all component build public API.
:Release Notes:
Do not export all the component build public API on CBE.

:Detailed Notes:
We found that CBE library was allocating a lot of memory for the ELF dynamic
loadable symbol table. It was because we were defining COMPONENT_BUILD on CBE
for exporting all the public symbols we would need out of libcbe.so (i.e. for
app_shell main, or webos webview).

But it was overkill, as webos webview and app runtime were already exported
no matter the value of COMPONENT_BUILD. So we just need to add the minimum
needed for app_shell (and eventually add the macros for the other CBE targets
as vanilla browser.

With this change libcbe.so goes from 96MB to 69MB. The number of exported
symbols go from 181k to 3k. Also, initial memory allocation went from 48MB to 26.

:Testing Performed:
Build finishes, libcbe.so is smaller now.

:QA Notes:
Test-cases: GT-26820

:Issues Addressed:
[PLAT-72851] [chromium68] Reduce disk size of chromium68

Change-Id: If795b612f9a42cc91eb646d63e8e202690bf4e24
Originally-Reviewed-on: http://gpro.lgsvl.com/228882
Reviewed-on: http://gpro.lgsvl.com/229004
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>