[open][neva_upstreamable][media] Fix video layout intrinsic size calculation
:Release Notes:
Prefer picking up of the widget view rectangle size over the natural size (if
the latter one is less), while calculating intrinsic size for a video layout.

:Detailed Notes:
Problem
-------
In legacy implementation, just the widget view rectangle size as well as video
rectangle in screen size (of a video element) values are compared (if NEVA web
media player is engaged). Thus, the video rectangle in screen size is returned
if it is greater than the widget view rectangle size, otherwise, the natural
size (got from the web media player) is returned instead. Unfortunately, that
leads to ill-formed treatment of corresponding video content dimensions within
the WebMediaPlayerNeva::UpdateVideoHoleBoundary() API, which (in turn) causes
improper scaling of a video streamed.

Solution
--------
Just to consider the widget view rectangle size when picking up the so called
natural size (from the web media player), and return the greatest one as the
intrinsic size for a video layout.

:Testing Performed:
1. Launch the Arirang application
2. Choose a video from the hot clip list
   (to switch to the fullscreen mode while playing the on-air video streaming)
3. Verify that the video is played in fullscreen mode

:QA Notes:
Test-cases: GT-22391, GT-21927

:Issues Addressed:
[PLAT-71439] [chromium68] Arirang Video doesn't play in full screen

Change-Id: Ie646b99794ab08d3568f60cbf314272aba57ebc3
Originally-Reviewed-on: http://gpro.lgsvl.com/228964
Reviewed-on: http://gpro.lgsvl.com/229003
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>