[open][neva_upstream][lttng] Finalize switching of the LTTng-UST support to dynamic linkage
:Release Notes:
Switched entire support of the LTTng-UST (Linux Tracing Toolkit: next generation
UserSpace Tracer) to be dynamically linked to once used.

:Detailed Notes:
Intro
-----
Currently, support of the LTTng-UST feature is designed for being engaged by
either generic Chromium stack (wherever correspoding alias tracepoints are set)
or by (from) injections machinery.

The one driven by the Chromium stack is dynamically linked to the LTTng-UST
external library, the other one - is statically linked (originally).

Problem
-------
There are two dedicated patchsets intended for providing support of the
LTTng-UST for either Chromium or injections already being successfully migrated
- http://gpro.lgsvl.com/218744 and http://gpro.lgsvl.com/220033 respectively.

Unfortunately, due to inaccurate migration of the latter patchset (related to
injections), zygote process is now crashed because of the LTTng-UST external
library is statically linked (since it spawns a listener thread on its startup,
which is treated as a severe violation by zygote).

Please, note that static linkage of the LTTng-UST support for injections turned
into a problem right after the USE_DYNAMIC_INJECTION_LOADING macro became
undefined. And since that feature is no more (moreover, looks like there are no
plans to revive it), forcing the LTTng-UST support for injections to be dyna-
mically linked seems as a must.

Solution
--------
To avoid the zygote assert checking (leading to its intentional termination), as
well as to preserve support of the LTTng-UST feature for both (Chromium stack
and injections machinery), and, finally, to avoid potential duplication (by
unintentional provision of two instances of the same library either dynamically
or statically linked), it seems reasonable to provide just a single entry point
(in terms of the LTTng-UST library coupling procedure - provider object) for
both of the deliverables at once.

Please, be informed that both of the LTTng-UST source patchsets are upstreamed
to the NEVA Configurable Framework already containing the suggested fix.

:Testing Performed:
1. Once the board is up, check whether the WAM app process tree is awaken:
   root@LGwebOS:~# ps axf | grep WebAppMgr

2. If awaken, stop the service:
   root@LGwebOS:~# systemctl stop webapp-mgr

3. Then restart manually the WAM binary:
   root@LGwebOS:~# sh -x /etc/systemd/system/scripts/webapp-mgr.sh

4. Verify that there is no more backtrace (with the zygote source code mentioned
   at the top) among the generated logging output

:QA Notes:
Test-cases: QR-1245
Depends on https://gpro.lgsvl.com/#/c/229290/

:Issues Addressed:
[PLAT-72894] [chromium68] Zygote process is crashed due to
             https://gpro.lgsvl.com/220033

Change-Id: I184d1db3200558de8eedd843adecaf0650284edb
Originally-Reviewed-on: http://gpro.lgsvl.com/229291
Reviewed-on: http://gpro.lgsvl.com/229695
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Andre Rosa <andre.rosa@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>