[open][neva_upstreamable][media] Fix Shaka Player "Bitcodin" shows black
:Release Notes:
Fix Shaka Player "Bitcodin" shows black

:Detailed Notes:
Problem:
Shaka Player with "Bitcodin" test uses adaptive streaming so the
resolution of video source is changed during playing video. When we set
video position to the platform video surface, we use both of
display_out_rect and video_source_rect, then media platform component maps
video_source_rect to display_out_rect. By adaptive streaming video source
resolution is changed and we update video position with the same
display_out_rect and changed video_source_rect. But media-pipeline seems
not to have decoded a new frame with changed resolution yet in this time.
If the resolution is changed from low to high, video displays black
because of this resolution miss matching. (if resolution is changed high
to low, video is zoomed out)

Solution:
To fix this issue, we don't update the video position if source_rect is
changed by changing video resolution. In this case platform handles it
without updating the video position.
Also when we calculate source_rect now we use video_size which was set
by platform media-pipeline. It is safer because this information is
shared with platform media component already.

:Testing Performed:
GT-21184 is tested on device and worked well.

:QA Notes:
Test-cases: GT-21184

:Issues Addressed:
[PLAT-73029] Black screen is displayed under "Bitcodin"
category in Shaka Player

Change-Id: I10e2ad78bea9f4b5a1888ded370ad7ffe86f9187
Originally-Reviewed-on: http://gpro.lgsvl.com/229386
Reviewed-on: http://gpro.lgsvl.com/229698
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>