[open][neva_upstreamable][ime] Clear selection text to correct prediction
:Release Notes:
Clear selection text in case editor is cleared

:Detailed Notes:
When we leave the page with edited input, input is cleared.
At that moment we need to clear the host view selected text,
which is used for predictions.

:Testing Performed:
Put word to User Guide search, then navigate back and
go to search again. Navigation back is tested with both
NRCU and mouse.

:QA Notes:
Test-cases: GT-22106

:Issues Addressed:
[PLAT-66339] Chromium 2018Q4 upgrade: analyze "ozone" patchset for Webkit
[PLAT-72892] VKB is displayed with the predictions
             when no inputs are given

Change-Id: Ia7104c730c98f7b930feeae8db177769fe4463b2
Originally-Reviewed-on: https://gpro.lgsvl.com/169927
Originally-Reviewed-on: http://gpro.lgsvl.com/229723
Reviewed-on: http://gpro.lgsvl.com/229903
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Andre Rosa <andre.rosa@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>