[open][neva_upstreamable][webrtc] Suspend audio capture if page is hidden
:Release Notes:
Suspend also audio capture on top of video capture if page is hidden.

:Detailed Notes:
On upstream chromium, for Android, the video capture is suspended if the
application is backgrounded or hidden. We took that code to also do the
same in webOS (adding the extra build guards to enable that code also
in our platform).

But, due to platform requests, we decided that, on WebOS, we should stop
the call and stop capturing video, but also audio.

This patch implements a parallel implementation to the video capture
suspend, but for audio capture.

:Testing Performed:
Launch skype, and do a call. Then while call is ongoing, switch to
live TV. When the app is backgrounded audio thread should not eat CPU
and sound should not be sent to the other peer.

Then, send Skype back to foreground. Stop the call and create a new
call. The new call audio and video capture should work.

:QA Notes:
Test-cases: N/A (this is being masked by the patch hanging the webRTC calls
on suspend).

:Issues Addressed:
[PLAT-45979] WebRTC: suspend audio capture if app is hidden
[PLAT-44883] webRTC: Audio Thread consuming CPU when app goes to background with active call
[PLAT-65882] Chromium 2018Q4 upgrade: migrate "webrtc" patchset

Change-Id: I4f4ad98ba7d1731d082833b05e7e7e7ca9ca09a0
Originally-Reviewed-on: http://gpro.lgsvl.com/183217
Originally-Reviewed-on: http://gpro.lgsvl.com/229894
Reviewed-on: http://gpro.lgsvl.com/230153
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>