[open][neva_upstreamable][media] Add use_unlimited_media_policy to allow application to override media policy
:Release Notes:
Introduce new webview API and renderer preference for managing media player.

:Detailed Notes:
The webview API name is SetUseUnlimitedMediaPolicy(bool). And
renderer preference name is use_unlimited_media_policy. This
API and preference change media state policy in a web application.
If a web application enable SetUseUnlimitedMediaPolicy as true,
all media players in the web application are managed as unlimited
media player, even if we use limited media policy such as
LimitedMediaActivationCountPolicy. We can use this API when we want
to relax strict policy in specific application scenario.

Another important point is, this API change behavior when more than
two web applications are foreground state(This situation seldom
occurs in webOS.). If one of web application SetUseUnlimitedMediaPolicy
as true, then media players in the application can't affect media
players in another web application. This means if we play a video
in the enabled application, any media players in another application
will not be suspended by policy. (Of course in app switching scenario,
the media player will be suspended as well.)

:Testing Performed:
1. Provide --max-activated-media-players=1 command line switch to
   web app manager. This will change media state policy.
2. Add an attribute "useUnlimitedMediaPolicy": true, in appinfo.json
3. Open the application and redirect to following test case:
   https://jira2.lgsvl.com/secure/attachment/469443/media_two_video_test.html
4. Play both video. Both video should not be suspended by playing
   another video(Only see playing state.).

:QA Notes:
Test-cases: GT-26199
Depends on https://gpro.lgsvl.com/232453

:Issues Addressed:
[PLAT-73639] while playing the music, black screen appeared in viewster video

Change-Id: I3384c3cee030a4b54c4bf59f873822d12d39de22
Originally-Reviewed-on: http://gpro.lgsvl.com/232196
Reviewed-on: http://gpro.lgsvl.com/232485
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>