fixup! [open][neva_upstreamable][ozone] Scroll the viewport when VKB is shown
:Release Notes:
Fix scrolling when VKB is activated but viewport is not moved

:Detailed Notes:
Scrolling did not work when we activate VKB without moving viewport
(click on input field from the upper part of the page).
It happened because there was a wrong variable that actually
stored viewport moving value on VKB activation. It was 0 when
VKB activated but viewport was not moved.
Changed it to real VKB height.

:Testing Performed:
1. Navigate to Settings -> Connection -> Wired Connection
2. Uncheck the 'Set Automatically' checkbox
3. Click in the 'IP' field
4. Verify that scrolling via mouse or MRCU unit is working

:QA Notes:
Test-cases: GT-22269

:Issues Addressed:
[PLAT-74901] The scrolling does not perform for the first two fields in the 'Settings' app

Change-Id: I2ab1c676b92ec2a33ab27103ab95e1b5336489d8
Originally-Reviewed-on: http://gpro.lgsvl.com/232752
Reviewed-on: http://gpro.lgsvl.com/233174
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>