[open][neva_upstreamable][app_runtime] Unset delegate on destroying WebAppWindow
:Release Notes:
Unset delegate on destroying WebAppWindow to prevent crashes.

:Detailed Notes:
On stress tests we detected a case where, on destroying
app_runtime::WebView while navigation was happening, then the
destruction of the NavigationHandle would notify DidFinishNavigation
to WebContents. But WebContents would then call the delegate, that
is the WebView being destroyed. That would crash as the virtual
that handles DidFinishNavigation was already freed.

The only way we had to reproduce this crash was using stress tests.

:Testing Performed:
Stress tests do not crash with the DidFinishNavigation stacktrace.

:QA Notes:
Test-cases: N/A (we could not find steps to reproduce yet, so only
  stress tests catch this).

:Issues Addressed:
[PLAT-74734] WebAppMgr crashes observed during stress test
[PLAT-75339] Stress tests crash: DidFinishNavigation

Change-Id: Iddcb0084350d84fe293545e80af9e4c52d7d85b7
Originally-Reviewed-on: http://gpro.lgsvl.com/232900
Reviewed-on: http://gpro.lgsvl.com/233175
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>