[open][neva_upstreamable][app_runtime] Fix updateViewportY crash
:Release Notes:
Fix updateViewportY crash

:Detailed Notes:
On stress tests we detected a case where, UpdateViewportY() crashed
in app_runtime::WebAppWindow while destruction of window was already
in progress. This patch adds null pointer checks to prevent the crash.

The only way we had to reproduce this crash was using stress tests.

:Testing Performed:
Stress tests do not crash with the DidFinishNavigation stacktrace.

:QA Notes:
Test-cases: N/A (we could not find steps to reproduce yet, so only
  stress tests catch this).

:Issues Addressed:
[PLAT-74734] WebAppMgr crashes observed during stress test
[PLAT-75338] Stress tests crash: UpdateViewportY

Change-Id: I2442d74c9bb9c80700670c4b2b022c79757040ab
Originally-Reviewed-on: http://gpro.lgsvl.com/232984
Reviewed-on: http://gpro.lgsvl.com/233173
Reviewed-by: OSE Helper Bot <ose@ci.lgsvl.net>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>